fix(db): cast bool params to i64 for BIGINT columns (Postgres strict typing)
Postgres refuses to coerce a boolean expression into a BIGINT column
('column X is of type bigint but expression is of type boolean'), where
SQLite silently stores 0/1. The schema stores booleans as BIGINT (the
SQLite convention), so the bool params bound to those columns must be
cast to i64 at the call site.

Fixed the four call sites the command smoke surfaced:
- db_create_calendar_event: all_day (bool)        -> all_day as i64
- db_update_calendar_event: all_day (Option<bool>) -> .map(|b| b as i64)
- db_create_calendar_calendar: visible (bool)      -> visible as i64
- db_update_calendar_calendar: visible (Option<bool>) -> .map(|b| b as i64)
- db_set_pattern_starred: starred (bool)          -> starred as i64
- db_set_word_starred: starred (bool)            -> starred as i64

(chat.rs and documents/crud.rs already used the 'as i64' / 'if b {1} else {0}'
pattern, so they were already portable.)

Also extended the command smoke to exercise calendar events (reserved-word
'start'/'end' columns) and translations (the daily_streaks.translations
streak upsert) — both backends now pass.

+++ app/core/src/db/patterns.rs
@@ -134,7 +134,7 @@ pub async fn db_set_pattern_starred(
     let db = db::conn(&conn)?;
     db.execute(
         "UPDATE patterns SET starred = ?1 WHERE id = ?2",
-        params![starred, pattern_id],
+        params![starred as i64, pattern_id],
     )
     .await
     .map_err(|e| e.to_string())?;
